+++ lang/en/local_boostnavigation.php
@@ -90,6 +90,17 @@ Further information to the parameters:
 </ul>
 Please note:
 <ul>
+<li>The link parameter can contain placeholders, for example /course/edit.php?id={courseid} to create a node linking to the current course\'s settings page. Placeholders are encapsulated in curly brackets and will be replaced automatically when then custom node is created.<br />Available placeholders are:
+<ul>
+<li>{courseid}: The course\'s (internal) ID</li>
+<li>{courseshortname}: The course\'s shortname</li>
+<li>{userid}: The logged in user\'s (internal) ID</li>
+<li>{userusername}: The logged in user\'s username</li>
+<li>{pagecontextid}: The current page\'s context ID</li>
+<li>{pagepath}: The current page\'s URL path</li>
+<li>{sesskey}: The sesskey to use in secured URLs</li>
+</ul>
+</li>
 <li>If the custom node does not show up in Boost\'s nav drawer, please check if all mandatory params are set correctly and if the optional language setting fits to your current Moodle user language.</li>
 <li>Due to the way how Boost\'s nav drawer is built in Moodle core, all custom nodes are displayed equally. Adding custom CSS classes, custom HTML element ids or a target attribute to open the link in a new window is impossible.</li>
 </ul>';
@@ -108,6 +119,17 @@ Further information to the parameters:
 Please note:
 <ul>
 <li>Pipe dividing for optional parameters is always needed if they are located between other options. This means that you have to separate params with the pipe character although they are empty. Also see the example for the Faculty of mathematics custom node above.</li>
+<li>The link parameter can contain placeholders, for example /course/edit.php?id={courseid} to create a node linking to the current course\'s settings page. Placeholders are encapsulated in curly brackets and will be replaced automatically when then custom node is created.<br />Available placeholders are:
+<ul>
+<li>{courseid}: The course\'s (internal) ID</li>
+<li>{courseshortname}: The course\'s shortname</li>
+<li>{userid}: The logged in user\'s (internal) ID</li>
+<li>{userusername}: The logged in user\'s username</li>
+<li>{pagecontextid}: The current page\'s context ID</li>
+<li>{pagepath}: The current page\'s URL path</li>
+<li>{sesskey}: The sesskey to use in secured URLs</li>
+</ul>
+</li>
 <li>If the custom node does not show up in Boost\'s nav drawer, please check if all mandatory params are set correctly, if the optional language setting fits to your current Moodle user language and if you are a member of the optional cohort setting.</li>
 <li>Due to the way how Boost\'s nav drawer is built in Moodle core, all custom nodes are displayed equally. Adding custom CSS classes, custom HTML element ids or a target attribute to open the link in a new window is impossible.</li>
 </ul>';